Add support for loading and saving input mappings
The key names for each InputMap action are saved in the user config file.
The controls UI now displays the current key bindings and saves the new value
when you change them; they are loading from the config on startup.

This uses new features introduced in and after Godot 1.1rc3, so the current
git version or 1.1stable are required.