Add stretch mode with preserved aspect ratio
This fixes an issue in the main menu where maximizing the window made
it possible to use other UI screens as they are next to each other
on the viewport.
The Camera2D in the level node has also been dropped, since both scenes
are now initiliased to double size with stretch mode.